requirements.txt
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.27.2
langchain==0.1.20
langchain-google-genai==1.0.3
google-generativeai==0.5.2
ollama==0.3.1
pydantic==2.9.2
langdetect==1.0.9

